+++ lib/ui/schedule/view_model/calendars/day_viewmodel.dart
@@ -1,7 +1,6 @@
 import 'package:calendar_view/calendar_view.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:notredame/ui/schedule/view_model/calendar_viewmodel.dart';
-import 'package:notredame/utils/utils.dart';
 
 class DayViewModel extends CalendarViewModel {
   DateTime daySelected = DateTime.now().withoutTime;
@@ -25,22 +24,21 @@ class DayViewModel extends CalendarViewModel {
   handleDateSelectedChanged(DateTime newDate) {
     daySelected = newDate;
 
-    // TODO don't remove everything
-    eventController.removeWhere((event) => true);
+    eventController.removeWhere((event) =>
+      event.date.withoutTime.difference(newDate.withoutTime).inDays.abs() > 1
+    );
 
-    // TODO don't store the events for 3 entire weeeks
-    List<CalendarEventData> eventsToAdd = selectedWeekCalendarEvents();
+    List<CalendarEventData> eventsToAdd = selectedDayCalendarEvents();
 
     eventController.addAll(eventsToAdd);
   }
 
-  List<CalendarEventData> selectedWeekCalendarEvents() {
+  List<CalendarEventData> selectedDayCalendarEvents() {
     final List<CalendarEventData> events = [];
 
-    final firstDayOfWeek = Utils.getFirstDayOfCurrentWeek(daySelected);
-    // We want to put events of previous week and next week in memory to make transitions smoother
-    for (int i = -7; i < 14; i++) {
-      final date = firstDayOfWeek.add(Duration(days: i));
+    // We want to put events of previous and next day in memory to make transitions smoother
+    for (int i = -1; i < 1; i++) {
+      final date = daySelected.add(Duration(days: i));
       final eventsForDay = calendarEventsFromDate(date);
       if (eventsForDay.isNotEmpty) {
         events.addAll(eventsForDay);
@@ -48,6 +46,4 @@ class DayViewModel extends CalendarViewModel {
     }
     return events;
   }
-
-  // TODO method to get number of events for each day in last week, current week and next week
 }
