+++ lib/ui/schedule/widgets/calendars/day_calendar.dart
@@ -1,38 +1,60 @@
-import 'package:calendar_view/calendar_view.dart';
+import 'package:calendar_view/calendar_view.dart' as calendar_view;
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:intl/intl.dart';
 import 'package:notredame/ui/core/themes/app_theme.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:notredame/ui/schedule/view_model/calendars/day_viewmodel.dart';
-import 'package:notredame/ui/schedule/widgets/day_view_header.dart';
 import 'package:notredame/ui/schedule/widgets/schedule_calendar_tile.dart';
 import 'package:stacked/stacked.dart';
 import 'package:notredame/data/services/signets-api/models/course_activity.dart';
 import 'package:notredame/ui/dashboard/widgets/course_activity_tile.dart';
+import 'package:table_calendar/table_calendar.dart';
+import 'package:notredame/ui/core/themes/app_palette.dart';
 
 class DayCalendar extends StatefulWidget {
-  final GlobalKey<DayViewState> dayViewKey;
+  final GlobalKey<calendar_view.DayViewState> dayViewKey;
+  final bool listView;
 
   const DayCalendar({
     super.key,
-    required this.dayViewKey
+    required this.dayViewKey,
+    required this.listView
   });
 
   @override
   State<DayCalendar> createState() => _DayCalendarState();
 }
 
-class _DayCalendarState extends State<DayCalendar> {
+class _DayCalendarState extends State<DayCalendar> with TickerProviderStateMixin {
+  late AnimationController _animationController;
+
+  @override
+  void initState() {
+    super.initState();
+    _animationController = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 400),
+    );
+
+    _animationController.forward();
+  }
+
+  @override
+  void dispose() {
+    _animationController.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder.reactive(
       viewModelBuilder: () => DayViewModel(intl: AppIntl.of(context)!),
       builder: (context, model, child) => Column(children: [
-        DayViewHeader(model: model, dayViewKey: widget.dayViewKey),
-        model.calendarViewSetting
-            ? _buildCalendar(model)
-            : _buildListView(model)
+        _dayViewHeader(model),
+        widget.listView
+            ? _buildListView(model)
+            : _buildCalendar(model)
       ]),
     );
   }
@@ -41,9 +63,9 @@ class _DayCalendarState extends State<DayCalendar> {
     final double heightPerMinute = (MediaQuery.of(context).size.height / 1200).clamp(0.45, 1.0);
 
     return Expanded(
-        child: DayView(
+        child: calendar_view.DayView(
             showVerticalLine: false,
-            dayTitleBuilder: DayHeader.hidden,
+            dayTitleBuilder: calendar_view.DayHeader.hidden,
             key: widget.dayViewKey,
             controller: model.eventController,
             onPageChange: (date, page) => ({
@@ -53,11 +75,11 @@ class _DayCalendarState extends State<DayCalendar> {
             }),
             backgroundColor: context.theme.scaffoldBackgroundColor,
             initialDay: model.daySelected,
-            hourIndicatorSettings: HourIndicatorSettings(
+            hourIndicatorSettings: calendar_view.HourIndicatorSettings(
               color: context.theme.appColors.scheduleLine,
             ),
             liveTimeIndicatorSettings:
-            LiveTimeIndicatorSettings(
+            calendar_view.LiveTimeIndicatorSettings(
               color: context.theme.textTheme.bodyMedium!.color!,
             ),
             heightPerMinute: heightPerMinute,
@@ -97,7 +119,7 @@ class _DayCalendarState extends State<DayCalendar> {
             const SizedBox(height: 6.0),
             _buildTitleForDate(model.daySelected),
             const SizedBox(height: 2.0),
-            if (model.selectedWeekCalendarEvents().isEmpty && !model.isBusy)
+            if (model.selectedDayCalendarEvents().isEmpty && !model.isBusy)
               Padding(
                 padding: const EdgeInsets.symmetric(vertical: 64.0),
                 child:
@@ -113,7 +135,7 @@ class _DayCalendarState extends State<DayCalendar> {
     );
   }
 
-  Widget _buildEventTile(List<CalendarEventData<dynamic>> events) {
+  Widget _buildEventTile(List<calendar_view.CalendarEventData<dynamic>> events) {
     if (events.isNotEmpty) {
       return ScheduleCalendarTile(
         title: events[0].title,
@@ -152,5 +174,131 @@ class _DayCalendarState extends State<DayCalendar> {
             ? const Divider(thickness: 1, indent: 30, endIndent: 30)
             : const SizedBox(),
         itemCount: events.length);
+  }
+
+  Widget _dayViewHeader(DayViewModel model) {
+    const Color selectedColor = AppPalette.etsLightRed;
+    final Color todayColor = context.theme.appColors.dayIndicatorWeekView;
+    final Color defaultColor = context.theme.appColors.scheduleLine;
+
+    return TableCalendar(
+      key: const Key("TableCalendar"),
+      locale: AppIntl.of(context)!.localeName,
+      selectedDayPredicate: (day) {
+        return isSameDay(model.daySelected, day);
+      },
+      headerStyle: HeaderStyle(
+          titleTextFormatter: (date, locale) =>
+              DateFormat.MMMMEEEEd(locale).format(model.daySelected),
+          titleCentered: true,
+          formatButtonVisible: false),
+      eventLoader: model.coursesActivitiesFor,
+      calendarFormat: CalendarFormat.week,
+      focusedDay: model.daySelected,
+      calendarBuilders: CalendarBuilders(
+          defaultBuilder: (context, date, _) =>
+              _buildHeaderDay(date, defaultColor, model, widget.dayViewKey),
+          outsideBuilder: (context, date, _) =>
+              _buildHeaderDay(date, defaultColor, model, widget.dayViewKey),
+          todayBuilder: (context, date, _) =>
+              _buildHeaderDay(date, todayColor, model, widget.dayViewKey),
+          selectedBuilder: (_, date, __) => FadeTransition(
+            opacity:
+            Tween(begin: 0.0, end: 1.0).animate(_animationController),
+            child: _buildHeaderDay(
+                date, selectedColor, model, widget.dayViewKey),
+          ),
+          markerBuilder: (context, date, events) {
+            final bool isSelected = isSameDay(date, model.daySelected);
+            final bool isToday = isSameDay(date, DateTime.now());
+            Color color = selectedColor;
+
+            if (!isSelected) {
+              if (isToday) {
+                color = todayColor;
+              } else {
+                color = defaultColor;
+              }
+            }
+
+            return _buildEventsMarker(model, date, events, color);
+          }),
+      // Those are now required by the package table_calendar ^3.0.0. In the doc,
+      // it is suggest to set them to values that won't affect user experience.
+      // Outside the range, the date are set to disable so no event can be loaded.
+      firstDay: DateTime.utc(2010, 12, 31),
+      lastDay: DateTime.utc(2100, 12, 31),
+    );
+  }
+
+  Widget _buildHeaderDay(DateTime date, Color color, DayViewModel model, GlobalKey<calendar_view.DayViewState> dayViewKey) =>
+      Container(
+        clipBehavior: Clip.antiAlias,
+        margin: const EdgeInsets.all(4.0),
+        decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(10),
+            border: Border.all(color: color)),
+        child: InkWell(
+          borderRadius: BorderRadius.circular(10),
+          onTap: () {
+            DateTime startingDate = model.daySelected;
+            setState(() {
+              model.handleDateSelectedChanged(date);
+            });
+
+            if (model.daySelected.difference(startingDate).inDays.abs() == 1) {
+              dayViewKey.currentState?.animateToDate(model.daySelected);
+            } else {
+              dayViewKey.currentState?.jumpToDate(model.daySelected);
+            }
+          },
+          child: Container(
+            padding: const EdgeInsets.only(top: 5.0, left: 6.0),
+            width: 100,
+            height: 100,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  '${date.day}',
+                  style: const TextStyle().copyWith(
+                    fontSize: 16.0,
+                    height: 1.2,
+                  ),
+                ),
+                if (date.month != DateTime.now().month ||
+                    date.year != DateTime.now().year)
+                  Text(DateFormat.MMM(AppIntl.of(context)!.localeName).format(date),
+                      style: const TextStyle(fontSize: 10.0)),
+              ],
+            ),
+          ),
+        ),
+      );
+
+  /// Build the square with the number of [events] for the [date]
+  Widget? _buildEventsMarker(
+      DayViewModel model, DateTime date, List events, Color color) {
+    if (events.isNotEmpty) {
+      return Positioned(
+        right: 1,
+        bottom: 1,
+        child: AnimatedContainer(
+          duration: const Duration(milliseconds: 200),
+          decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(5), color: color),
+          width: 16.0,
+          height: 16.0,
+          child: Text(
+            '${events.length}',
+            textAlign: TextAlign.center,
+            style: const TextStyle().copyWith(
+              fontSize: 12.0,
+            ),
+          ),
+        ),
+      );
+    }
+    return null;
   }
 }
